Add files via upload
This SVI INTER-VLAN is configured using a 3560 layer 3 switch which works as a 
switch, router and also a DHCP server which means it is capable of multitasking and 
such can be perfect for small or a medium size business who wants to save cost and 
still keep an active and modern device for their network environment. It has a layer 
2 switch which can also convert into a layer 3 interface to serve purpose of routing. 
5 VLANs will be created in this Lab which will consist of a server VLAN and four 
departments which consist of two computer.